% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{as.data.frame.movetrack}
\alias{as.data.frame.movetrack}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{movetrack}(x, ...)
}
\arguments{
\item{x}{An object of class \code{movetrack}.}

\item{...}{Unused; for compatibility with the generic method.}
}
\description{
Coerce a \code{movetrack} object to a \code{data.frame}.
}
\seealso{
\code{summary.movetrack()}
}
