% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{as.data.frame.stantrackr}
\alias{as.data.frame.stantrackr}
\title{Coerce a \code{stantrackr} object to a Data Frame}
\usage{
\method{as.data.frame}{stantrackr}(x)
}
\arguments{
\item{x}{An object of class \code{stantrackr}.}
}
\description{
Coerce a \code{stantrackr} object to a Data Frame
}
