% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{distance}
\alias{distance}
\title{Calculate distances}
\usage{
distance(x)
}
\arguments{
\item{x}{: \code{data.frame}. Data in wide format}
}
\value{
\code{vector}
}
\description{
Calculate distances
}
\details{
Returns distances between consecutive locations in meters.
}
