% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localisation.R
\name{est_loc}
\alias{est_loc}
\title{Estimate location}
\usage{
est_loc(data = NULL, det_range = 12000, dtime = 1)
}
\arguments{
\item{data}{: \code{data.frame}. Motus data}

\item{det_range}{: \code{numeric}. Detection range of antennas in meters}

\item{dtime}{: \code{numeric}. Time interval in minutes}
}
\value{
\code{data.frame}
}
\description{
Estimate location
}
\details{
\itemize{
\item Estimate location for each detection: half of detection range
along the directional beam
\item Derive oscillating measurement error
\item Weighted means (by signal strength) for each minute interval
}
}
