% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Extract estimates}
\usage{
estimates(fit, mdata = NULL, prob = 0.9)
}
\arguments{
\item{fit}{A \code{CmdStanMCMC} object containing the model output.}

\item{mdata}{A \code{data.frame} containing metadata for each time interval.}

\item{prob}{A scalar \verb{[0, 1]} specifying the mass within the credible
interval.}
}
\value{
Returns a \code{data.frame} containing the mean and lower and upper bound of the
highest density interval for each time interval.
}
\description{
Extract estimates from model output.
}
