% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{getDraws}
\alias{getDraws}
\title{Extract draws}
\usage{
getDraws(fit, nsim = 50)
}
\arguments{
\item{fit}{An object of class \code{stantrackr}.}

\item{nsim}{The number of simulations to extract; defaults to \code{50}.}
}
\value{
A \code{data.frame} with the draws.
}
\description{
Extract draws from a \code{stantrackr} object.
}
