% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Estimate location}
\usage{
locate(
  data = NULL,
  ID = "tagDeployID",
  ts = "ts",
  sig = "sig",
  aLon = "recvDeployLon",
  aLat = "recvDeployLat",
  aBearing = "antBearing",
  det_range = 12,
  dtime = 2
)
}
\arguments{
\item{data}{A \code{data.frame} containing the telemetry data.}

\item{ID}{Unique identifier for individuals or tag deployments.}

\item{ts}{Timestamp column.}

\item{sig}{Signal strength column.}

\item{aLon}{Antenna longitude column.}

\item{aLat}{Antenna latitude column.}

\item{aBearing}{Antenna bearing column.}

\item{det_range}{Assumed maximum detection range of antennas in kilometres.}

\item{dtime}{Time interval in minutes for which point estimates are to be
calculated.}
}
\value{
Returns a \code{data.frame} containing estimated coordinates and measurement
errors for each time interval together with the proportions of time
intervals \code{w}.
}
\description{
Calculate point estimates based on antenna bearing and signal strength.
}
\details{
This function performs the following steps as described in
\href{https://doi.org/10.1016/j.ecolmodel.2018.08.006}{Baldwin et al. 2018}:
\itemize{
\item Estimate locations for each detection: half of the maximum detection
range \code{det_range} along the directional beam.
\item Derive oscillating measurement error arising from antenna geometry and
orientation.
\item Calculate weighted means (by signal strength) for each time interval
\code{dtime}.
}
}
