% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{mapTrack}
\alias{mapTrack}
\title{Map model result}
\usage{
mapTrack(fit, id = NULL, nsim = 50, lwd = 2, alpha = 0.1)
}
\arguments{
\item{fit}{An object of class \code{stantrackr}.}

\item{id}{The individuals to plot. Defaults to \code{NULL} which plots all
individuals.}

\item{nsim}{The number of posterior draws to plot. Defaults to \code{50}.}

\item{lwd}{The line width for the mean trajectory. Defaults to \code{2}.}

\item{alpha}{The alpha value for the posterior draws. Defaults to \code{0.1}.}
}
\value{
Returns an overview map with the mean trajectories and \code{nsim} posterior draws
per individual.
}
\description{
Map individual flight trajectories and model uncertainty.
}
\examples{
\dontrun{
mapTrack(fit)
mapTrack(fit, nsim = 100, alpha = 0.05)
}

}
