% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localisation.R
\name{model_loc}
\alias{model_loc}
\title{Model data}
\usage{
model_loc(data = NULL, refresh = 100)
}
\arguments{
\item{data}{: \code{data.frame}. Data}

\item{refresh}{: \code{numeric}. refresh value for Stan}
}
\value{
\code{data.frame}
}
\description{
Model data
}
\details{
returns location estimates for each time interval
}
