% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\docType{data}
\name{motusData}
\alias{motusData}
\title{Motus test data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4464 rows and 7 columns.
}
\usage{
motusData
}
\description{
Motus test data
}
\references{
\url{https://motus.org/}
}
\keyword{data}
