% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim}
\alias{optim}
\title{Optimisation method for \code{track()}}
\usage{
optim(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the point estimate data.}

\item{...}{Additional arguments passed to \code{cmdstanr::optimize()}.}
}
\description{
Model flight path from point estimates using the optimisation method from
\code{cmdstanr}.
}
