% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot.stantrackr}
\alias{plot.stantrackr}
\title{Plot model results}
\usage{
\method{plot}{stantrackr}(x, vars = c("lon", "lat"), id = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{stantrackr}.}

\item{vars}{The variables to plot. Defaults to \code{c("lon", "lat")}.}

\item{id}{The individuals to plot. Defaults to \code{NULL} which plots all
individuals.}

\item{...}{Additional arguments passed to \code{stantrckr::summary()}.}
}
\value{
Returns one or multiple \code{ggplot} plots.
}
\description{
Plot model results per individual and variable.
}
\examples{
\dontrun{
# Set ggplot theme
theme_set(theme_bw(base_size = 20))

# Plot
plot(fit)
plot(fit, vars = "speed", prob = 0.89, ci = "ETI")
}

}
\seealso{
\code{stantrackr::summary()}
}
