% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.movetrack}
\alias{print.movetrack}
\title{Print}
\usage{
\method{print}{movetrack}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{movetrack}.}

\item{digits}{The minimal number of \emph{significant} digits; defaults to \code{3}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print a summary for a \code{movetrack} object.
}
\seealso{
\code{summary.movetrack()}
}
