% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.stantrackr}
\alias{print.stantrackr}
\title{Print a summary for a \code{stantrackr} object}
\usage{
\method{print}{stantrackr}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{stantrackr}.}

\item{digits}{The minimal number of \emph{significant} digits; defaults to \code{3}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\description{
Print a summary for a \code{stantrackr} object
}
\seealso{
\code{summary.stantrackr()}
}
