% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{speed}
\alias{speed}
\title{Calculate speed}
\usage{
speed(x)
}
\arguments{
\item{x}{A \code{data.frame} in wide format.}
}
\value{
Returns a vector of speeds between consecutive locations in m/s.
}
\description{
Calculate flight speed between consecutive locations.
}
