% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{distance}
\alias{distance}
\title{Calculate distances}
\usage{
distance(x)
}
\arguments{
\item{x}{A \code{data.frame} in wide format.}
}
\value{
Returns a vector of distances between consecutive locations in metres.
}
\description{
Calculate distances between consecutive locations.
}
