% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localisation.R
\name{estimates}
\alias{estimates}
\title{Extract estimates}
\usage{
estimates(fit, mdata = NULL, prob = 0.9)
}
\arguments{
\item{fit}{: \code{stanfit}. Model output}

\item{mdata}{: \code{data.frame}. Metadata for each time interval}

\item{prob}{: \code{numeric}. Probability for HDI}
}
\value{
\code{data.frame}
}
\description{
Extract estimates
}
\details{
Returns mean, lower and upper bound of the highest density interval.
}
