% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Estimate location}
\usage{
locate(data = NULL, det_range = 12, dtime = 1)
}
\arguments{
\item{data}{\code{data.frame} Motus data}

\item{det_range}{\code{numeric} Detection range of antennas in kilometers}

\item{dtime}{\code{numeric} Time interval in minutes}
}
\value{
a \code{data.frame}
}
\description{
Estimate location based on antenna bearing and signal strength.
}
\details{
\itemize{
\item Estimate location for each detection: half of detection range
along the directional beam
\item Derive oscillating measurement error
\item Weighted means (by signal strength) for each minute interval
}
}
