% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Estimate locations}
\usage{
locate(
  data,
  ID = "tagDeployID",
  ts = "ts",
  sig = "sig",
  aLon = "recvDeployLon",
  aLat = "recvDeployLat",
  aType = NULL,
  aBearing = "antBearing",
  aRange = 12,
  dTime = 2
)
}
\arguments{
\item{data}{A \code{data.frame} containing the telemetry data.}

\item{ID}{Unique identifier for individuals or tag deployments.}

\item{ts}{Timestamp column.}

\item{sig}{Signal strength column.}

\item{aLon}{Antenna longitude column.}

\item{aLat}{Antenna latitude column.}

\item{aType}{Antenna type column, only required for antenna-specific
detection ranges.}

\item{aBearing}{Antenna bearing column.}

\item{aRange}{Assumed maximum detection range of antennas in kilometres.
Can be a single value or a named list of values for different antenna
types.}

\item{dTime}{Time interval in minutes for which point estimates are to be
calculated.}
}
\value{
Returns a \code{data.frame} containing estimated coordinates and measurement
errors for each time interval together with the proportions of time
intervals \code{w}.
}
\description{
Calculate point estimates based on antenna bearing and signal strength.
}
\details{
This function performs the following steps as described in
\href{https://doi.org/10.1016/j.ecolmodel.2018.08.006}{Baldwin et al. 2018}:
\itemize{
\item Estimate locations for each detection: half of the maximum detection
range \code{aRange} along the directional beam.
\item Derive oscillating measurement error arising from antenna geometry and
orientation.
\item Calculate weighted means (by signal strength) for each time interval
\code{dTime}.
}
}
\examples{
\dontrun{
data(motusData)
locate(motusData)
locate(motusData, dTime = 1, aRange = 10)
locate(motusData, aType = "antType", aRange = list("yagi-5"=10, "yagi-6"=12))
}

}
