% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim}
\alias{optim}
\title{Optimisation method for \code{track()}}
\usage{
optim(data, output_dir = getwd(), ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the point estimate data.}

\item{output_dir}{The directory to save the \code{CmdStanMCMC} objects to;
defaults to the current working directory. If \code{NULL}, the objects are not
saved.}

\item{...}{Additional arguments passed to \code{cmdstanr::optimize()}.}
}
\description{
Model flight path from point estimates using the optimisation method from
\code{cmdstanr}.
}
