% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localisation.R
\name{speed}
\alias{speed}
\title{Calculate speed}
\usage{
speed(x)
}
\arguments{
\item{x}{: \code{data.frame}. Data in wide format}
}
\value{
\code{vector}
}
\description{
Calculate speed
}
\details{
Returns speed between consecutive locations in m/s.
}
