% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.stantrackr}
\alias{summary.stantrackr}
\title{Create a summary of a \code{stantrackr} object}
\usage{
\method{summary}{stantrackr}(object, var = "lon", ci = "HDI", prob = 0.9)
}
\arguments{
\item{object}{An object of class \code{stantrackr}.}

\item{ci}{The method used to calculate the credible intervals. Available
options are \code{'HDI'} for the highest posterior density interval and \code{'ETI'}
for the equal-tailed interval; defaults to \code{'HDI'}.}

\item{prob}{The probability mass of the credible interval; defaults to \code{0.9}.}
}
\description{
Create a summary of a \code{stantrackr} object
}
