% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Model data}
\usage{
track(data = NULL, refresh = 100)
}
\arguments{
\item{data}{\code{data.frame} Data}

\item{refresh}{\code{numeric} Refresh value for Stan}
}
\value{
a \code{data.frame}
}
\description{
Model data using a DCRW model.
}
\details{
Returns location estimates for each time interval.
}
