% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Model data}
\usage{
track(data, output_dir = getwd(), ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the point estimate data.}

\item{output_dir}{The directory to save the \code{CmdStanMCMC} objects to;
defaults to the current working directory. If \code{NULL}, the objects are not
saved.}

\item{...}{Additional arguments passed to \link[cmdstanr]{sample}.}
}
\value{
Returns a \code{stantrackr} object containing the posterior distributions for
longitude, latitude, distance and speed.
}
\description{
Model flight path from point estimates using a DCRW model.
}
\details{
This function is a wrapper around \code{cmdstanr::sample()} and uses a difference
correlated random walk model (DCRW) to estimate individual flight paths. The
model is described in more detail in
\href{https://doi.org/10.1890/04-1852}{Jonsen et al. 2005} and
\href{https://doi.org/10.1016/j.ecolmodel.2018.08.006}{Baldwin et al. 2018}. To
learn more about state-space models in animal movement in general,
\href{https://doi.org/10.1002/ecm.1470}{Auger-Méthé et al. 2021} is a good
starting point.
}
