% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localisation.R
\name{track}
\alias{track}
\title{Model data}
\usage{
track(data = NULL, refresh = 100)
}
\arguments{
\item{data}{: \code{data.frame}. Data}

\item{refresh}{: \code{numeric}. Refresh value for Stan}
}
\value{
\code{data.frame}
}
\description{
Model data
}
\details{
Returns location estimates for each time interval.
}
