% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Model data}
\usage{
track(data, states = 1, method = "mcmc", ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the point estimate data.}

\item{states}{The number of states to use in the model; defaults to \code{1}.}

\item{method}{The estimation method to use. Either \code{"mcmc"} or \code{"optim"};
defaults to \code{"mcmc"}.}

\item{...}{Additional arguments passed to \code{cmdstanr::sample()} or
\code{cmdstanr::optimize()}, respectively.}
}
\value{
Returns a \code{data.frame} containing estimates for longitude, latitude, distance
and speed per time interval or a \code{stantrackr} object including the full
posterior distributions.
}
\description{
Model flight path from point estimates using a DCRW model.
}
\details{
This function calls \href{https://mc-stan.org/}{Stan} via
\href{https://mc-stan.org/cmdstanr/index.html}{cmdstanr} and uses a difference
correlated random walk model (DCRW) to estimate individual flight paths.
The model is described in more detail in
\href{https://doi.org/10.1890/04-1852}{Jonsen et al. 2005} and
\href{https://doi.org/10.1016/j.ecolmodel.2018.08.006}{Baldwin et al. 2018}. To
learn more about state-space models in animal movement in general,
\href{https://doi.org/10.1002/ecm.1470}{Auger-Méthé et al. 2021} is a good
starting point. Available estimation methods are Stan's main Markov chain
Monte Carlo algorithm and Stan's optimizer to obtain a (penalized) maximum
likelihood estimate or a maximum a posteriori estimate (if \code{jacobian=TRUE}).
See the \href{https://mc-stan.org/docs/cmdstan-guide}{CmdStan User's Guide} for
more details.
}
\examples{
\dontrun{
# Load data
data(motusData)

# Estimate locations
loc <- locate(motusData, dtime = 2)

# Model flight paths
track(loc, states = 2, parallel_chains = 4)
track(loc, "optim", refresh = 1e3)
}

}
\seealso{
\code{cmdstanr::sample()} \code{cmdstanr::optimize()}
}
